requirements.txt
aiogram
loguru
python-dotenv
requests
matplotlib
